scraper/requirements.txt
Scrapy==2.11.2
supabase==2.5.0
python-dotenv==1.0.1
scrapy-rotating-proxies==0.6.2
itemadapter==0.9.0
w3lib==2.2.1
